[ALPS04967689] SEPolicy: Add neverallow rule for proc
[Detail]
Do not allow access to the generic proc label. This is too broad.
Instead, if access to part of proc is desired, it should have a
more specific label.
TODO: Remove mtk_hal_audio/audioserver and so on once there are no violations.
EX.
   r_dir_file(hal_audio, proc)
   hal_server_domain(mtk_hal_audio, hal_audio)
   hal_client_domain(audioserver, hal_audio)

[Solution]
1.Add neverallow rule for proc.
2.Remove the conflicting SEPolicies.

Change-Id: Ie932149f8c642d4a05152117f1166daeaf9b2cff
CR-Id: ALPS04967689
Feature: [Android Default] SELinux, SEAndroid, and SE-MTK