[ALPS04967419] SEPolicy: Add neverallow rule for sysfs
[Detail]
Do not allow access to the generic sysfs label. This is too broad.
Instead, if access to part of sysfs is desired, it should have a
more specific label.
TODO: Remove hal_usb/mtk_hal_usb and so on once there are no violations.
EX.
   allow hal_usb sysfs:file write;
   hal_server_domain(mtk_hal_usb, hal_usb)

   r_dir_file(hal_wifi, sysfs_type)
   hal_server_domain(mtk_hal_wifi, hal_wifi)

[Solution]
1.Add neverallow rule for sysfs.
2.Remove the conflicting SEPolicies.

MTK-Commit-Id: 86296cf74da59aa881bb2ae8ad868195b67079d5

Change-Id: I304a1a87b23623e320ff7346da9d10a09264152b
CR-Id: ALPS04967419
Feature: [Android Default] SELinux, SEAndroid, and SE-MTK