Put in SELinux holes for other cameras
Right now, anyone using CameraGo is told to fully disable
SELinux for the entire device in order to be able to save
pictures they take.  This is not an issue with CameraGo,
as OpenCamera has the same behavior.

With this CL, we weaken and put holes in the device SELinux
policy until we're able to save a picture with OpenCamera.
This is a much better state than having folks disable
SELinux altogether, and also should make it much easier to
test with this camera.

Bug: 152082918
Test: OpenCamera app is able to store pictures and videos without disabling selinux for the device.
Change-Id: I14afc331c12dd0316e8c90cc878374e0a1ee94f6